requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
playwright>=1.44.0
